filter is now generic component + fix
* fix select and combobox maxCharBtn
* Now filter is a component extending default fields.
For the moment, we have 2 types of filters: select and keyword.
We have default values that avoid filter ("all" for select and "" for keyword).
Filters are fields so we need to provide a "field" key with same structure as a Field.
We have to define "keys" that will be the keys the filter value will check.
We can set filter :"default" in order to filter the base keys of an object.
We can set filter :"settings" in order to filter settings, data must be an advanced template.