chore: Update headers.lua to include interest-cohort in Permissions-Policy header the proper way
This commit modifies the headers.lua file to include the "interest-cohort" directive in the "Permissions-Policy" header. This change ensures that the "interest-cohort" directive is added to the header when the "DISABLE_FLOC" variable is set to "yes". By doing so, the application adheres to the necessary privacy settings and improves the handling of user data.