refactor: Remove unused UI files and update Dockerfile
The Dockerfile for the Linux environment has been updated to remove the unnecessary copying of UI files, specifically the "static" and "templates" directories. These directories are no longer used in the application and can be safely removed. This change reduces the size of the Docker image and improves build times.