feat: Update SSL cipher preference for TLSv1.3 + Also if the TLSv1.3 isn't part of the configured SSL_PROTOCOLS, then HTTP3 is disabled to avoid configuration errors
This commit updates the SSL cipher preference for TLSv1.3 in the default server HTTP configuration and the server HTTP and server stream SSL certificate Lua configurations. When TLSv1.3 is the only one used, the `ssl_prefer_server_ciphers` directive is set to "off" to disable the preference for server ciphers. This ensures compatibility with the latest TLSv1.3 standards.